refactor: приведение структуры проекта к Java-стандартам
