+++ src/main/java/ru/aston/hw1/animal/Animal.java
@@ -1,0 +1,6 @@
+package main.java.ru.aston.hw1.animal;
+
+abstract class Animal {
+
+    abstract String getSpecies();
+}
