+++ src/main/java/ru/aston/hw1/animal/Waterfowl.java
@@ -1,0 +1,6 @@
+package main.java.ru.aston.hw1.animal;
+
+public interface Waterfowl {
+
+    boolean liveInWater();
+}
