+++ src/main/java/ru/aston/hw1/animal/Whale.java
@@ -1,0 +1,14 @@
+package main.java.ru.aston.hw1.animal;
+
+public class Whale extends Mammal implements Waterfowl {
+
+    @Override
+    public boolean liveInWater() {
+        return true;
+    }
+
+    @Override
+    public String getSpecies() {
+        return "Cetacea";
+    }
+}
