+++ src/main/java/ru/aston/hw1/vehicle/CanCarryCargo.java
@@ -1,0 +1,8 @@
+package main.java.ru.aston.hw1.vehicle;
+
+public interface CanCarryCargo {
+
+    void loadCargo();
+
+    void unloadCargo();
+}
