+++ src/main/java/ru/aston/hw1/vehicle/HasPropeller.java
@@ -1,0 +1,6 @@
+package main.java.ru.aston.hw1.vehicle;
+
+public interface HasPropeller {
+
+    boolean isStart();
+}
