+++ src/main/java/ru/aston/hw1/vehicle/HasWings.java
@@ -1,0 +1,6 @@
+package main.java.ru.aston.hw1.vehicle;
+
+public interface HasWings {
+
+    String wingType();
+}
