+++ src/main/java/ru/aston/hw1/vehicle/Vehicle.java
@@ -1,0 +1,6 @@
+package main.java.ru.aston.hw1.vehicle;
+
+abstract class Vehicle {
+
+    abstract void move();
+}
