+++ src/main/java/ru/aston/hw3/FileOperationException.java
@@ -1,0 +1,12 @@
+package main.java.ru.aston.hw3;
+
+
+/**
+ * Настраиваемое исключения при работе с файлами.
+ */
+public class FileOperationException extends Exception {
+
+    public FileOperationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
