+++ src/main/resources/ru/aston/hw3/example-file.txt
