--- src/ru/aston/hw1/animal/Animal.java
@@ -1,6 +1,0 @@
-package ru.aston.hw1.animal;
-
-abstract class Animal {
-
-    abstract String getSpecies();
-}
