--- src/ru/aston/hw1/animal/Bear.java
@@ -1,14 +1,0 @@
-package ru.aston.hw1.animal;
-
-public class Bear extends Mammal implements FurBearing {
-
-    @Override
-    public boolean hasFur() {
-        return true;
-    }
-
-    @Override
-    public String getSpecies() {
-        return "Ursus";
-    }
-}
