--- src/ru/aston/hw1/animal/FurBearing.java
@@ -1,6 +1,0 @@
-package ru.aston.hw1.animal;
-
-public interface FurBearing {
-
-    boolean hasFur();
-}
