--- src/ru/aston/hw1/animal/Mammal.java
@@ -1,12 +1,0 @@
-package ru.aston.hw1.animal;
-
-abstract class Mammal extends Animal implements Vertebrate {
-
-    @Override
-    abstract String getSpecies();
-
-    @Override
-    public boolean hasSpine() {
-        return true;
-    }
-}
