--- src/ru/aston/hw1/animal/Vertebrate.java
@@ -1,6 +1,0 @@
-package ru.aston.hw1.animal;
-
-public interface Vertebrate {
-
-    boolean hasSpine();
-}
