--- src/ru/aston/hw1/animal/Waterfowl.java
@@ -1,6 +1,0 @@
-package ru.aston.hw1.animal;
-
-public interface Waterfowl {
-
-    boolean liveInWater();
-}
