--- src/ru/aston/hw1/vehicle/CanCarryCargo.java
@@ -1,8 +1,0 @@
-package ru.aston.hw1.vehicle;
-
-public interface CanCarryCargo {
-
-    void loadCargo();
-
-    void unloadCargo();
-}
