--- src/ru/aston/hw1/vehicle/HasPropeller.java
@@ -1,6 +1,0 @@
-package ru.aston.hw1.vehicle;
-
-public interface HasPropeller {
-
-    boolean isStart();
-}
