--- src/ru/aston/hw1/vehicle/HasWheels.java
@@ -1,6 +1,0 @@
-package ru.aston.hw1.vehicle;
-
-public interface HasWheels {
-
-    int getWheelCount();
-}
