--- src/ru/aston/hw1/vehicle/HasWings.java
@@ -1,6 +1,0 @@
-package ru.aston.hw1.vehicle;
-
-public interface HasWings {
-
-    String wingType();
-}
