--- src/ru/aston/hw1/vehicle/Vehicle.java
@@ -1,6 +1,0 @@
-package ru.aston.hw1.vehicle;
-
-abstract class Vehicle {
-
-    abstract void move();
-}
