--- src/ru/aston/hw3/FileOperationException.java
@@ -1,12 +1,0 @@
-package ru.aston.hw3;
-
-
-/**
- * Настраиваемое исключения при работе с файлами.
- */
-public class FileOperationException extends Exception {
-
-    public FileOperationException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
